Use flush for all ldiff changes to an openldap_access resource
Flushes happen every time a resource changes, whether it is creation,
destroy or modify. Ensure we do not needlesly copy-paste the ldapmodify
code in those three separate instances.

This also ensures that if a new property is ever added, we do not need
create a separate setter.

This commit thus also fixes a bug where if a `what` of an
openldap_access resource changed, it would not trigger an update.
The setter for that property was lost in the refactor of commit
91e972303d91634c8c0b3e8a99ea8867d0de7836. -> Fixes #420.